refactor(tui/thread): use dynamic import for subcommand
