+++ packages/opencode/src/cli/cmd/tui/thread/util.ts
@@ -1,0 +1,8 @@
+import path from "path"
+import { Filesystem } from "@/util/filesystem"
+
+export function resolveThreadDirectory(project?: string, envPWD = process.env.PWD, cwd = process.cwd()) {
+  const root = Filesystem.resolve(envPWD ?? cwd)
+  if (project) return Filesystem.resolve(path.isAbsolute(project) ? project : path.join(root, project))
+  return Filesystem.resolve(cwd)
+}
