+++ packages/opencode/src/index.ts
@@ -23,7 +23,7 @@ import { GithubCommand } from "./cli/cmd/github/command"
 import { ExportCommand } from "./cli/cmd/export/command"
 import { ImportCommand } from "./cli/cmd/import/command"
 import { AttachCommand } from "./cli/cmd/tui/attach/command"
-import { TuiThreadCommand } from "./cli/cmd/tui/thread"
+import { TuiThreadCommand } from "./cli/cmd/tui/thread/command"
 import { AcpCommand } from "./cli/cmd/acp/command"
 import { EOL } from "os"
 import { WebCommand } from "./cli/cmd/web"
