+++ packages/opencode/src/temporary.ts
@@ -1,5 +1,5 @@
 import yargs from "yargs"
-import { TuiThreadCommand } from "./cli/cmd/tui/thread"
+import { TuiThreadCommand } from "./cli/cmd/tui/thread/command"
 import { InstallationVersion } from "@opencode-ai/core/installation/version"
 import { hideBin } from "yargs/helpers"
 import { Log } from "./node"
